Cast boolean fields to integer and standardize casing in SQL queries for consistency and improved readability.
Query 021 to 050

+++ postgres_sql_queries/query_038.sql
@@ -9,7 +9,7 @@
 SELECT		logicallabel_id, 
 				version, 
 				label, 
-				mutable, 
+				mutable::integer, 
 				labeltype, 
 				TO_CHAR(datecreated, 'YYYY-MM-DD HH24:MI:SS') as datecreated, 
 			    TO_CHAR(datelastmodified, 'YYYY-MM-DD HH24:MI:SS') as datelastmodified,
